requirements.txt
PySide6>=6.5
python-vlc>=3.0.181
